# docs/swagger.yaml
basePath: /v1
definitions:
...
  entity.Level:
    properties:
      code:
        example: A1
        type: string
      createdAt:
        example: "2026-08-06T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      languageId:
        example: 1
        type: integer
      translations:
        items:
          $ref: '#/definitions/entity.LevelTranslation'
        type: array
    type: object
  entity.LevelTranslation:
    properties:
      languageId:
        example: 1
        type: integer
      name:
        example: Beginner
        type: string
    type: object
  entity.Topic:
    properties:
      createdAt:
        example: "2026-08-06T00:00:00Z"
        type: string
      iconUrl:
        example: https://example.com/travel.svg
        type: string
      id:
        example: 1
        type: integer
      isActive:
        example: true
        type: boolean
      slug:
        example: travel
        type: string
      translations:
        items:
          $ref: '#/definitions/entity.TopicTranslation'
        type: array
    type: object
  entity.TopicTranslation:
    properties:
      languageId:
        example: 1
        type: integer
      name:
        example: Travel
        type: string
    type: object
...
  github_com_evrone_go-clean-template_internal_controller_restapi_v1_admin_request.LevelTranslation:
    properties:
      languageId:
        example: 1
        type: integer
      name:
        example: Beginner
        maxLength: 100
        type: string
    required:
    - languageId
    - name
    type: object
...
  github_com_evrone_go-clean-template_internal_controller_restapi_v1_admin_request.SaveLevel:
    properties:
      code:
        example: A1
        maxLength: 20
        type: string
      languageId:
        example: 1
        type: integer
      translations:
        items:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_admin_request.LevelTranslation'
        minItems: 1
        type: array
    required:
    - code
    - languageId
    - translations
    type: object
  github_com_evrone_go-clean-template_internal_controller_restapi_v1_admin_request.SaveTopic:
    properties:
      iconUrl:
        example: https://example.com/travel.svg
        type: string
      isActive:
        example: true
        type: boolean
      slug:
        example: travel
        maxLength: 100
        type: string
      translations:
        items:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_admin_request.TopicTranslation'
        minItems: 1
        type: array
    required:
    - isActive
    - slug
    - translations
    type: object
...
  github_com_evrone_go-clean-template_internal_controller_restapi_v1_admin_request.TopicStatus:
    properties:
      isActive:
        example: true
        type: boolean
    required:
    - isActive
    type: object
  github_com_evrone_go-clean-template_internal_controller_restapi_v1_admin_request.TopicTranslation:
    properties:
      languageId:
        example: 1
        type: integer
      name:
        example: Travel
        maxLength: 100
        type: string
    required:
    - languageId
    - name
    type: object
...
  internal_controller_restapi_v1_admin.errorDoc:
    properties:
      error:
        example: message
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: User API with Firebase authentication and role-based access control
  title: Go Clean Template API
  version: "1.0"
paths:
...
  /admin/levels:
    get:
      description: Return levels, optionally filtered by their source language
      operationId: admin-list-levels
      parameters:
      - description: Source language ID
        in: query
        name: language_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Level'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
      security:
      - BearerAuth: []
      summary: List levels
      tags:
      - admin-levels
    post:
      consumes:
      - application/json
      description: Create a level with its multilingual translations
      operationId: admin-create-level
      parameters:
      - description: Level and translations
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_admin_request.SaveLevel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Level'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
      security:
      - BearerAuth: []
      summary: Create level
      tags:
      - admin-levels
  /admin/levels/{id}:
    delete:
      description: Delete a level only when no video references it
      operationId: admin-delete-level
      parameters:
      - description: Level ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
      security:
      - BearerAuth: []
      summary: Delete level
      tags:
      - admin-levels
    put:
      consumes:
      - application/json
      description: Replace a level and its multilingual translations
      operationId: admin-update-level
      parameters:
      - description: Level ID
        in: path
        name: id
        required: true
        type: integer
      - description: Level and translations
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_admin_request.SaveLevel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Level'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
      security:
      - BearerAuth: []
      summary: Update level
      tags:
      - admin-levels
...
  /admin/topics:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Topic'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
      security:
      - BearerAuth: []
      summary: List topics
      tags:
      - admin-topics
    post:
      consumes:
      - application/json
      parameters:
      - description: Topic
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_admin_request.SaveTopic'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Topic'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
      security:
      - BearerAuth: []
      summary: Create topic with translations
      tags:
      - admin-topics
  /admin/topics/{id}:
    delete:
      description: Delete a topic and its translations only when no video references
        it
      parameters:
      - description: Topic ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
      security:
      - BearerAuth: []
      summary: Delete topic
      tags:
      - admin-topics
    put:
      consumes:
      - application/json
      parameters:
      - description: Topic ID
        in: path
        name: id
        required: true
        type: integer
      - description: Topic
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_admin_request.SaveTopic'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Topic'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
      security:
      - BearerAuth: []
      summary: Update topic and translations
      tags:
      - admin-topics
  /admin/topics/{id}/status:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Topic ID
        in: path
        name: id
        required: true
        type: integer
      - description: Visibility
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_admin_request.TopicStatus'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Topic'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
      security:
      - BearerAuth: []
      summary: Enable or disable topic visibility
      tags:
      - admin-topics
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"